Refine KDA state routing and tests
