+++ tests/attention/test_request_state_manager.py
@@ -1,26 +1,14 @@
 # SPDX-License-Identifier: Apache-2.0
 from __future__ import annotations
 
-from types import SimpleNamespace
-
 import mlx.core as mx
-import mlx.nn as nn
 import numpy as np
 import pytest
 
-from tests.stub_runner import make_bailing_hybrid_plan, make_gdn_hybrid_plan
+from tests.stub_runner import make_gdn_hybrid_plan
 from vllm_metal.attention.caches.gdn_cache import GDNPagedStateCache
-from vllm_metal.attention.context import (
-    PagedAttentionContext,
-    clear_context,
-    set_context,
-)
-from vllm_metal.attention.impls.kda import KDAPagedAttentionWrapper
-from vllm_metal.attention.impls.mla import MLAPagedAttentionWrapper
-from vllm_metal.attention.runtime.hybrid import (
-    HybridPagedAttentionRuntime,
-    MLAHybridPagedAttentionRuntime,
-)
+from vllm_metal.attention.context import PagedAttentionContext
+from vllm_metal.attention.runtime.hybrid import HybridPagedAttentionRuntime
 from vllm_metal.attention.state import RequestStateManager
 
 
@@ -40,17 +28,6 @@ def _make_cache(*, num_layers: int = 2, max_seqs: int = 2) -> GDNPagedStateCache
 
 def _make_context() -> PagedAttentionContext:
     return PagedAttentionContext(slot_mapping=[])
-
-
-def _make_bailing_runtime(num_layers: int) -> MLAHybridPagedAttentionRuntime:
-    return MLAHybridPagedAttentionRuntime(
-        hybrid_plan=make_bailing_hybrid_plan(num_layers),
-        max_num_seqs=2,
-        num_kv_heads=1,
-        head_dim=6,
-        block_size=4,
-        dtype=mx.float32,
-    )
 
 
 class TestRequestStateManager:
@@ -289,127 +266,3 @@ class TestHybridPagedAttentionRuntime:
         assert not cache.has_pending_conv_state(0)
         assert not cache.has_pending_recurrent_state(0)
         assert runtime.state_manager.needs_materialize is False
-
-
-class _FakeBailingKDA(nn.Module):
-    q_proj = object()
-    k_proj = object()
-    v_proj = object()
-    q_conv1d = object()
-    k_conv1d = object()
-    v_conv1d = object()
-    projection_size = 4
-    conv_kernel_size = 2
-
-
-class _FakeBailingMLA(nn.Module):
-    pass
-
-
-class _FakeBailingLayer:
-    def __init__(self, attention: nn.Module) -> None:
-        self.attention = attention
-
-
-class TestKDAPagedAttentionWrapper:
-    def test_multi_request_prefill_and_decode_match_mlx_lm(self) -> None:
-        from mlx_lm.models.bailing_moe_v3 import BailingKDA, ModelArgs
-        from mlx_lm.models.cache import ArraysCache
-
-        mx.random.seed(17)
-        args = ModelArgs(
-            hidden_size=16,
-            num_attention_heads=2,
-            head_dim=4,
-            short_conv_kernel_size=3,
-        )
-        inner = BailingKDA(args)
-        projection_size = args.num_attention_heads * args.head_dim
-        state_cache = GDNPagedStateCache(
-            num_layers=1,
-            max_seqs=2,
-            conv_kernel_dim=args.short_conv_kernel_size,
-            conv_dim=3 * projection_size,
-            num_v_heads=args.num_attention_heads,
-            value_head_dim=args.head_dim,
-            key_head_dim=args.head_dim,
-            initial_seqs=2,
-            dtype=mx.float32,
-        )
-        wrapper = KDAPagedAttentionWrapper(inner, 0, 0, state_cache)
-        references = [ArraysCache(size=4), ArraysCache(size=4)]
-
-        for cu_seqlens, grouped_slots in (([0, 3, 5], False), ([0, 1, 2], True)):
-            x = mx.random.normal((1, cu_seqlens[-1], args.hidden_size)).astype(
-                mx.float32
-            )
-            expected = mx.concatenate(
-                [
-                    inner(x[:, start:end], cache=reference)
-                    for start, end, reference in zip(
-                        cu_seqlens[:-1], cu_seqlens[1:], references, strict=True
-                    )
-                ],
-                axis=1,
-            )
-
-            set_context(
-                PagedAttentionContext(
-                    slot_mapping=[],
-                    cu_seqlens=cu_seqlens,
-                    state_slot_mapping=None if grouped_slots else [0, 1],
-                    state_group_slot_mappings=([0, 1],) if grouped_slots else None,
-                )
-            )
-            try:
-                actual = wrapper(x)
-            finally:
-                clear_context()
-
-            mx.eval(actual, expected)
-            np.testing.assert_allclose(
-                np.array(actual), np.array(expected), rtol=1e-5, atol=1e-5
-            )
-
-        expected_conv = mx.concatenate(
-            [mx.concatenate(reference.cache[:3], axis=-1) for reference in references]
-        )
-        expected_recurrent = mx.concatenate([reference[3] for reference in references])
-        np.testing.assert_array_equal(
-            np.array(state_cache.conv_states[0]), np.array(expected_conv)
-        )
-        np.testing.assert_array_equal(
-            np.array(state_cache.recurrent_states[0]), np.array(expected_recurrent)
-        )
-
-
-class TestMLAHybridPagedAttentionRuntime:
-    def test_patches_interleaved_layers_with_compact_cache_indices(self) -> None:
-        runtime = _make_bailing_runtime(4)
-        runtime.initialize(num_blocks=3)
-        model = SimpleNamespace(
-            model=SimpleNamespace(
-                layers=[
-                    _FakeBailingLayer(_FakeBailingKDA()),
-                    _FakeBailingLayer(_FakeBailingMLA()),
-                    _FakeBailingLayer(_FakeBailingKDA()),
-                    _FakeBailingLayer(_FakeBailingMLA()),
-                ]
-            )
-        )
-
-        assert runtime.patch_model(model) == 4
-
-        layers = model.model.layers
-        assert all(
-            isinstance(layers[idx].attention, KDAPagedAttentionWrapper)
-            for idx in (0, 2)
-        )
-        assert all(
-            isinstance(layers[idx].attention, MLAPagedAttentionWrapper)
-            for idx in (1, 3)
-        )
-        assert [layers[idx].attention._kda_cache_idx for idx in (0, 2)] == [0, 1]
-        assert [layers[idx].attention._mla_layer_idx for idx in (1, 3)] == [0, 1]
-        assert runtime._cache.num_layers == 2
-        assert runtime.state_cache.num_layers == 2
